Move Tab component
