+++ src/docs/pages/TabsPage.tsx
@@ -1,7 +1,7 @@
 import type { FC } from 'react';
 import { HiAdjustments, HiClipboardList, HiUserCircle } from 'react-icons/hi';
 import { MdDashboard } from 'react-icons/md';
-import { Tabs } from '../../lib';
+import { Tabs } from '../../lib/components/molecules';
 import type { CodeExample } from './DemoPage';
 import { DemoPage } from './DemoPage';
