+++ src/lib/components/atoms/DarkThemeToggle/DarkThemeToggle.spec.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { describe, expect, it } from 'vitest';
 import { DarkThemeToggle } from '.';
-import { HelloInternet } from '../bosons/HelloInternet';
+import { HelloInternet } from '../../bosons/HelloInternet';
 
 describe.concurrent('Dark theme toggle', () => {
   it('should toggle the theme when `Space` is pressed', async () => {
