+++ src/lib/components/atoms/index.ts
@@ -1,5 +1,6 @@
 export * from './Alert';
 export * from './Button';
+export * from './DarkThemeToggle';
 export * from './Progress';
 export * from './Section';
 export * from './Toast';
