+++ src/lib/components/bosons/HelloInternet/HelloInternetTheme.ts
@@ -1,6 +1,7 @@
 import { DeepPartial } from '../..';
 import { AlertTheme } from '../../atoms/Alert/Alert.theme';
 import type { ButtonTheme } from '../../atoms/Button/Button.theme';
+import { DarkThemeToggleTheme } from '../../atoms/DarkThemeToggle/DarkThemeToggle.theme';
 import type { ProgressColors, ProgressSizes } from '../../atoms/Progress';
 import type { SectionTheme } from '../../atoms/Section/Section.theme';
 import { ToastTheme } from '../../atoms/Toast/Toast.theme';
@@ -22,7 +23,7 @@ import { ModalTheme } from '../../molecules/Modal/Modal.theme';
 import type { StarSizes } from '../../Rating';
 import type { SidebarCTAColors } from '../../Sidebar/SidebarCTA';
 import type { SpinnerColors, SpinnerSizes } from '../../Spinner';
-import type { TabStyleItem, TabStyles } from '../../Tab';
+import { TabsTheme } from '../../molecules/Tab/Tabs.theme';
 
 export type CustomHelloInternetTheme = DeepPartial<HelloInternetTheme>;
 
@@ -139,10 +140,7 @@ export interface HelloInternetTheme extends Record<string, unknown> {
       snap: string;
     };
   };
-  darkThemeToggle: {
-    base: string;
-    icon: string;
-  };
+  darkThemeToggle: DarkThemeToggleTheme,
   footer: {
     base: string;
     container: string;
@@ -398,19 +396,7 @@ export interface HelloInternetTheme extends Record<string, unknown> {
     };
     size: SpinnerSizes;
   };
-  tab: {
-    base: string;
-    tablist: {
-      base: string;
-      styles: TabStyles;
-      tabitem: {
-        base: string;
-        styles: TabStyleItem<TabStyles>;
-        icon: string;
-      };
-    };
-    tabpanel: string;
-  };
+  tab: TabsTheme;
   toast: ToastTheme;
   tooltip: TooltipTheme;
   dropdown: {
