+++ src/lib/components/index.ts
@@ -7,7 +7,6 @@ export * from './Breadcrumb';
 export * from './ButtonGroup';
 export * from './Card';
 export * from './Carousel';
-export * from './DarkThemeToggle';
 export * from './Dropdown';
 export * from './Footer';
 export * from './FormControls';
@@ -17,7 +16,6 @@ export * from './Pagination';
 export * from './Rating';
 export * from './Sidebar';
 export * from './Spinner';
-export * from './Tab';
 export * from './Table';
 export * from './Timeline';
