+++ src/lib/components/molecules/Tab/index.tsx
@@ -1,9 +1,9 @@
 import classNames from 'classnames';
 import type { ComponentProps, FC, KeyboardEvent, PropsWithChildren, ReactElement } from 'react';
 import { Children, useEffect, useId, useMemo, useRef, useState } from 'react';
-import { excludeClassName } from '../../helpers/exclude';
-import type { FlowbiteBoolean } from '../bosons/HelloInternet/HelloInternetTheme';
-import { useTheme } from '../bosons/HelloInternet/ThemeContext';
+import { excludeClassName } from '../../../helpers/exclude';
+import type { FlowbiteBoolean } from '../../bosons/HelloInternet/HelloInternetTheme';
+import { useTheme } from '../../bosons/HelloInternet/ThemeContext';
 import type { TabProps } from './TabItem';
 import { TabItem } from './TabItem';
