+++ src/lib/components/molecules/index.tsx
@@ -1,1 +1,2 @@
 export * from './Modal';
+export * from './Tab';
